+++ packages/common/src/types/dbt.ts
@@ -1,0 +1,78 @@
+import { DimensionType, MetricType } from './field';
+
+export enum SupportedDbtAdapter {
+    BIGQUERY = 'bigquery',
+    DATABRICKS = 'databricks',
+    SNOWFLAKE = 'snowflake',
+    REDSHIFT = 'redshift',
+    POSTGRES = 'postgres',
+}
+
+export type DbtNode = {
+    unique_id: string;
+    resource_type: string;
+};
+export type DbtRawModelNode = DbtNode & {
+    columns: { [name: string]: DbtModelColumn };
+    config?: { meta?: DbtModelMetadata };
+    meta: DbtModelMetadata;
+    database: string | null;
+    schema: string;
+    name: string;
+    tags: string[];
+    relation_name: string;
+    depends_on: DbtTableDependency;
+    description?: string;
+    root_path: string;
+    patch_path: string | null;
+};
+export type DbtModelNode = DbtRawModelNode & {
+    database: string;
+};
+type DbtTableDependency = {
+    nodes: string[];
+};
+export type DbtModelColumn = {
+    name: string;
+    description?: string;
+    meta: DbtColumnMetadata;
+    data_type?: DimensionType;
+};
+
+type DbtModelMetadata = DbtModelLightdashConfig & {};
+
+type DbtModelLightdashConfig = {
+    label?: string;
+    joins?: DbtModelJoin[];
+};
+type DbtModelJoin = {
+    join: string;
+    sql_on: string;
+};
+type DbtColumnMetadata = DbtColumnLightdashConfig & {};
+type DbtColumnLightdashConfig = {
+    dimension?: DbtColumnLightdashDimension;
+    metrics?: { [metricName: string]: DbtColumnLightdashMetric };
+};
+
+type DbtColumnLightdashDimension = {
+    name?: string;
+    label?: string;
+    type?: DimensionType;
+    description?: string;
+    sql?: string;
+    time_intervals?: string | string[];
+    hidden?: boolean;
+    round?: number;
+    format?: string;
+};
+
+export type DbtColumnLightdashMetric = {
+    label?: string;
+    type: MetricType;
+    description?: string;
+    sql?: string;
+    hidden?: boolean;
+    round?: number;
+    format?: string;
+};