+++ packages/frontend/src/components/Explorer.tsx
@@ -5,6 +5,7 @@ import {
     Classes,
     Collapse,
     Dialog,
+    Divider,
     H5,
     Menu,
     MenuItem,
@@ -32,6 +33,7 @@ import { useQueryResults } from '../hooks/useQueryResults';
 import {
     useAddVersionMutation,
     useDeleteMutation,
+    useDuplicateMutation,
     useSavedQuery,
     useUpdateMutation,
 } from '../hooks/useSavedQuery';
@@ -103,6 +105,8 @@ export const Explorer: FC<Props> = ({ savedQueryUuid }) => {
     const [sqlIsOpen, setSqlIsOpen] = useState<boolean>(false);
     const [vizIsOpen, setVizisOpen] = useState<boolean>(!!savedQueryUuid);
     const totalActiveFilters: number = countTotalFilterRules(filters);
+    const chartId = savedQueryUuid || '';
+    const { mutate: duplicateChart } = useDuplicateMutation(chartId);
     const [activeVizTab, setActiveVizTab] = useState<ChartType>(
         ChartType.CARTESIAN,
     );
@@ -399,13 +403,30 @@ export const Explorer: FC<Props> = ({ savedQueryUuid }) => {
                                             content={
                                                 <Menu>
                                                     <MenuItem
-                                                        icon="add"
-                                                        text="Save chart as"
-                                                        onClick={() =>
-                                                            setIsQueryModalOpen(
-                                                                true,
-                                                            )
+                                                        icon={
+                                                            hasUnsavedChanges()
+                                                                ? 'duplicate'
+                                                                : 'add'
                                                         }
+                                                        text={
+                                                            hasUnsavedChanges()
+                                                                ? 'Duplicate'
+                                                                : 'Save as new chart'
+                                                        }
+                                                        onClick={() => {
+                                                            if (
+                                                                savedQueryUuid &&
+                                                                hasUnsavedChanges()
+                                                            ) {
+                                                                duplicateChart(
+                                                                    chartId,
+                                                                );
+                                                            } else {
+                                                                setIsQueryModalOpen(
+                                                                    true,
+                                                                );
+                                                            }
+                                                        }}
                                                     />
                                                     <MenuItem
                                                         icon="control"
@@ -416,10 +437,10 @@ export const Explorer: FC<Props> = ({ savedQueryUuid }) => {
                                                             )
                                                         }
                                                     />
-
+                                                    <Divider />
                                                     <MenuItem
-                                                        icon="delete"
-                                                        text="Delete chart"
+                                                        icon="trash"
+                                                        text="Delete"
                                                         intent="danger"
                                                         onClick={() => {
                                                             getDashboards(