+++ packages/frontend/src/components/Home/LatestSavedCharts/index.tsx
@@ -1,11 +1,16 @@
 import { AnchorButton } from '@blueprintjs/core';
-import { SessionUser } from 'common';
+import { SessionUser, SpaceQuery } from 'common';
 import React, { FC } from 'react';
+import {
+    useDeleteMutation,
+    useUpdateMutation,
+} from '../../../hooks/useSavedQuery';
 import { useSavedCharts } from '../../../hooks/useSpaces';
 import { useTimeAgo } from '../../../hooks/useTimeAgo';
+import ActionCardList from '../../common/ActionCardList';
+import SavedQueryForm from '../../SavedQueries/SavedQueryForm';
 import LatestCard from '../LatestCard';
 import {
-    ChartName,
     CreateChartButton,
     UpdatedLabel,
     ViewAllButton,
@@ -36,6 +41,13 @@ export const UpdatedInfo: FC<{
 const LatestSavedCharts: FC<{ projectUuid: string }> = ({ projectUuid }) => {
     const savedChartsRequest = useSavedCharts(projectUuid);
     const savedCharts = savedChartsRequest.data || [];
+    const featuredCharts = savedCharts
+        .sort(
+            (a, b) =>
+                new Date(b.updatedAt).getTime() -
+                new Date(a.updatedAt).getTime(),
+        )
+        .slice(0, 5);
     return (
         <LatestCard
             isLoading={savedChartsRequest.isLoading}
@@ -60,27 +72,20 @@ const LatestSavedCharts: FC<{ projectUuid: string }> = ({ projectUuid }) => {
                 )
             }
         >
-            {savedCharts
-                .sort(
-                    (a, b) =>
-                        new Date(b.updatedAt).getTime() -
-                        new Date(a.updatedAt).getTime(),
-                )
-                .slice(0, 5)
-                .map(({ uuid, name, updatedAt, updatedByUser }) => (
-                    <ChartName
-                        key={uuid}
-                        minimal
-                        href={`/projects/${projectUuid}/saved/${uuid}`}
-                        alignText="left"
-                    >
-                        {name}
-                        <UpdatedInfo
-                            updatedAt={updatedAt}
-                            user={updatedByUser}
-                        />
-                    </ChartName>
-                ))}
+            <ActionCardList
+                title="Saved charts"
+                useUpdate={useUpdateMutation}
+                useDelete={useDeleteMutation()}
+                dataList={featuredCharts}
+                getURL={(savedQuery: SpaceQuery) => {
+                    const { uuid } = savedQuery;
+                    return `/projects/${projectUuid}/saved/${uuid}`;
+                }}
+                ModalContent={SavedQueryForm}
+                isHomePage
+                isChart
+            />
+
             {savedCharts.length === 0 && (
                 <CreateChartButton
                     minimal