+++ packages/frontend/src/components/common/modal/ModalActionButtons.tsx
@@ -1,6 +1,8 @@
-import { Button, Menu, MenuItem } from '@blueprintjs/core';
+import { Button, Divider, Menu, MenuItem } from '@blueprintjs/core';
 import { Popover2 } from '@blueprintjs/popover2';
-import React, { Dispatch, SetStateAction, useState } from 'react';
+import React, { Dispatch, SetStateAction, useEffect, useState } from 'react';
+import { useDuplicateDashboardMutation } from '../../../hooks/dashboard/useDashboard';
+import { useDuplicateMutation } from '../../../hooks/useSavedQuery';
 import { ActionTypeModal } from './ActionModal';
 
 type ModalActionButtonsProps = {
@@ -9,14 +11,25 @@ type ModalActionButtonsProps = {
     setActionState: Dispatch<
         SetStateAction<{ actionType: number; data?: any }>
     >;
+    isChart?: boolean;
 };
 
 const ModalActionButtons = ({
     data,
     url,
     setActionState,
+    isChart,
 }: ModalActionButtonsProps) => {
     const [isOpen, setIsOpen] = useState(false);
+    const [itemId, setItemId] = useState<string>('');
+    const { mutate: duplicateChart } = useDuplicateMutation(itemId);
+    const { mutate: duplicateDashboard } =
+        useDuplicateDashboardMutation(itemId);
+    const isDashboardPage = url.includes('/dashboards');
+
+    useEffect(() => {
+        setItemId(data.uuid);
+    }, []);
 
     return (
         <Popover2
@@ -42,7 +55,40 @@ const ModalActionButtons = ({
                     />
                     <MenuItem
                         role="button"
-                        icon="delete"
+                        icon="duplicate"
+                        text="Duplicate"
+                        onClick={(e) => {
+                            e.preventDefault();
+                            e.stopPropagation();
+                            if (!isChart) {
+                                duplicateDashboard(itemId);
+                            }
+                            if (isChart) {
+                                duplicateChart(itemId);
+                            }
+                            setIsOpen(false);
+                        }}
+                    />
+                    {!isDashboardPage && (
+                        <MenuItem
+                            icon="insert"
+                            text="Add to Dashboard"
+                            onClick={(e) => {
+                                e.preventDefault();
+                                e.stopPropagation();
+                                setIsOpen(false);
+                                setActionState({
+                                    actionType:
+                                        ActionTypeModal.ADD_TO_DASHBOARD,
+                                    data,
+                                });
+                            }}
+                        />
+                    )}
+                    <Divider />
+                    <MenuItem
+                        role="button"
+                        icon="trash"
                         text="Delete"
                         intent="danger"
                         onClick={(e) => {
@@ -70,6 +116,10 @@ const ModalActionButtons = ({
             />
         </Popover2>
     );
+};
+
+ModalActionButtons.defaultProps = {
+    isChart: false,
 };
 
 export default ModalActionButtons;