+++ packages/frontend/src/providers/ExplorerProvider.tsx
@@ -1,5 +1,7 @@
 import {
+    AdditionalMetric,
     FieldId,
+    Metric,
     MetricQuery,
     SortField,
     TableCalculation,
@@ -30,6 +32,7 @@ export enum ActionType {
     UPDATE_TABLE_CALCULATION,
     DELETE_TABLE_CALCULATION,
     RESET_SORTING,
+    SET_ADDITIONAL_METRICS,
 }
 
 type Action =
@@ -73,6 +76,10 @@ type Action =
       }
     | {
           type: ActionType.RESET_SORTING;
+      }
+    | {
+          type: ActionType.SET_ADDITIONAL_METRICS;
+          payload: Metric[];
       };
 
 interface ExplorerReduceState {
@@ -87,6 +94,7 @@ interface ExplorerReduceState {
     columnOrder: string[];
     limit: number;
     tableCalculations: TableCalculation[];
+    additionalMetrics: AdditionalMetric[] | undefined;
 }
 
 interface ExplorerState extends ExplorerReduceState {
@@ -134,6 +142,7 @@ const defaultState: ExplorerReduceState = {
     limit: 500,
     tableCalculations: [],
     selectedTableCalculations: [],
+    additionalMetrics: [],
 };
 
 const calcColumnOrder = (
@@ -294,6 +303,12 @@ function reducer(
                 filters: action.payload,
             };
         }
+        case ActionType.SET_ADDITIONAL_METRICS: {
+            return {
+                ...state,
+                additionalMetrics: action.payload,
+            };
+        }
         case ActionType.SET_COLUMN_ORDER: {
             return {
                 ...state,
@@ -361,6 +376,7 @@ function reducer(
                 ]),
             };
         }
+
         default: {
             throw new Error(`Unhandled action type`);
         }